requirements.txt
numpy
pandas
matplotlib
scikit-learn
python-telegram-bot